fix: complete Oxigraph 0.5.x migration -- lib and bin now compile (96 -> 0 errors)
Third pass on the Oxigraph 0.5.x migration. `cargo build` and `cargo build --bins`
now succeed; `./target/debug/spreadsheet-mcp --help` runs.

Oxigraph 0.5.x API migration:
- Store::quads_for_pattern now takes concrete Ref types (NamedOrBlankNodeRef /
  NamedNodeRef / TermRef / GraphNameRef) instead of `impl Into<...>`. Rewrote every
  call site in graph_integrity.rs to pass `.as_ref()` / `.as_ref().into()`, and
  replaced `&*rdf::TYPE` with the `rdf::TYPE` NamedNodeRef constant.
- Graph::iter() now yields TripleRef: graph_validator.rs match arms use
  TermRef/NamedOrBlankNodeRef variants, `.into_owned()` instead of `.to_owned()`.
- shacl.rs: replaced the ambiguous `S: Into<SubjectRef<'a>>` generic on the
  get_object/get_*_value helpers with a concrete NamedOrBlankNodeRef parameter
  (fixes 10 E0283 inference failures); Store::contains takes QuadRef by value;
  Literal::datatype() already returns a ref; RdfFormat::JsonLd is now a struct
  variant; `unwrap_or(&format!(..))` -> `cloned().unwrap_or_else(|| format!(..))`.
- QueryResults now borrows the Store: SafeQueryResult gained a lifetime parameter
  rather than claiming 'static.
- QuerySolution is no longer Deserialize. QueryResultCache::get already stored a
  lossy JSON projection, so its return type is now the honest `Vec<CachedSolution>`
  (serde_json::Map) instead of a Vec<QuerySolution> that could never round-trip.

Dependency correction:
- prettyplease 0.3 requires syn ^3 while this crate pins syn 2.0, which produced a
  syn::file::File vs syn::File mismatch. Pinned prettyplease to 0.2 (verified
  against crates.io: prettyplease 0.3.0 depends on syn ^3).

Other real fixes:
- logging.rs: take the concrete SDK Tracer returned by install_batch instead of the
  global BoxedTracer, which does not implement PreSampledTracer.
- MappingError gained From<ValidationError> so query_wrappers' `?` works.
- Added missing Serialize/Deserialize/Clone/JsonSchema derives across the jira_*
  and ggen_sync response types actually required by their containers.
- ggen_sync AtomicWriter::commit was still using HashMap methods on a Vec.
- src/generated/mcp_tools.rs: the generating ggen template still emits `#[tool(...)]`
  on free functions taking `&SpreadsheetServer`; rmcp 0.11's macro boxes a 'static
  future, so that form cannot compile. Attributes stripped with an explanatory
  header note -- the live router is the `#[tool_router] impl` in src/server.rs and
  these functions have no callers. The real fix belongs in the ggen template.

Also fixed 6 errors in the (still-failing) lib test target: a `mod tests` name
collision with the real `pub mod tests` submodule, DodCheck imported from the wrong
module, a stale `name()` override, and three missing test-scope imports.

Verification: `cargo build` clean (135 warnings, 0 errors); binary runs.
The `--tests` targets still do not compile (separate backlog, ~44 errors, largely
tests referencing a `ggen_mcp` crate name and stale struct shapes).

+++ src/sparql/performance.rs
@@ -296,7 +296,7 @@ impl QueryAnalyzer {
         let has_specific_predicates = query.contains("rdfs:label") || query.contains("rdf:type");
         let has_literals = query.contains('\"');
 
-        let mut selectivity: f32 = 0.3; // Base selectivity
+        let mut selectivity: f64 = 0.3; // Base selectivity
 
         if has_filters {
             selectivity += 0.2;
